validation des entité et des web services

+++ src/main/java/com/projetb32/koulouwakel/service/PictureService.java
@@ -4,7 +4,13 @@ import com.projetb32.koulouwakel.entity.Picture;
 
 import com.projetb32.koulouwakel.repository.PictureRepository;
 import org.springframework.stereotype.Component;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.nio.file.StandardCopyOption;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -13,13 +19,27 @@ import java.util.Optional;
 public class PictureService {
     private final PictureRepository  pictureRepository;
 
+    private static String uploadDirectory =System.getProperty("user.dir")+"/src/uploads";
+
     public PictureService(PictureRepository pictureRepository) {
         super();
         this.pictureRepository = pictureRepository;
     }
 
-    public Picture addPicture(Picture picture) {
-        return pictureRepository.save(picture);
+    public Picture addPicture(MultipartFile file)
+    {
+        Picture p = new Picture();
+        p.setLabel(file.getOriginalFilename());
+        Path fileNameAndPath = Paths.get(uploadDirectory,file.getOriginalFilename());
+        StringBuilder fileName = new StringBuilder();
+        fileName.append(file.getOriginalFilename());
+        try {
+            Files.write(fileNameAndPath, file.getBytes());
+        } catch (IOException e){
+            e.printStackTrace();
+        }
+
+        return pictureRepository.save(p);
     }
 
     public List<Picture> getAllPicture() {
@@ -29,6 +49,8 @@ public class PictureService {
     }
 
     public Optional<Picture> getPictureById(Long id) {
+
+
 
         return pictureRepository.findById(id);
 
@@ -41,19 +63,35 @@ public class PictureService {
 
     public void deletePicture(Long id) {
 
-        pictureRepository.deleteById(id);
+        Picture picture = pictureRepository.findById(id).get();
+        if(picture != null)
+        {
+            Path fileNameAndPath = Paths.get(uploadDirectory,picture.getLabel());
+            try {
+                Files.delete(fileNameAndPath);
+                pictureRepository.deleteById(id);
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
 
-    }
+        }
 
-    public Picture updatePicture(Long id,Picture picture) {
+    public Picture updatePicture(Long id,String name) {
 
+        Picture picture = pictureRepository.findById(id).get();
+        Path fileNameAndPathSource = Paths.get(uploadDirectory,picture.getLabel());
+        picture.setLabel(name);
+        Path  currentlyFileNameAndPath = Paths.get(uploadDirectory,name);
 
-        Picture pictureFound = pictureRepository.findById(id).get();
+        try {
+            Files.move(fileNameAndPathSource,currentlyFileNameAndPath,StandardCopyOption.REPLACE_EXISTING);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
 
+        pictureRepository.save(picture);
 
-        pictureFound.setLabel(picture.getLabel());
-        pictureRepository.save(pictureFound);
-
-        return pictureFound;
+        return picture;
     }
 }
